Prepare googletest framework: move bioinform tests to subdir of tests, add example_gtest files and example tests, modifications in-progress to CMakeLists.txt, tests/CMakeLists.txt
