+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "extern/googletest"]
+	path = extern/googletest
+	url = ../../google/googletest.git
