+++ src/example_gtest.cpp
@@ -1,0 +1,16 @@
+#include "example_gtest.h"
+
+double add_numbers(const double f1, const double f2)
+{
+    return f1 + f2;
+}
+
+double subtract_numbers(const double f1, const double f2)
+{
+    return f1 - f2;
+}
+
+double multiply_numbers(const double f1, const double f2)
+{
+    return f1 * f2;
+}
