+++ src/example_gtest.h
@@ -1,0 +1,7 @@
+#pragma once
+
+double add_numbers(const double f1, const double f2);
+
+double subtract_numbers(const double f1, const double f2);
+
+double multiply_numbers(const double f1, const double f2);
