+++ tests/bioinform_tests/core_run.sh
