+++ tests/bioinform_tests/create_bioinform_test.sh
