+++ tests/bioinform_tests/gold_standard_results/bioinform_long_goldstandard.tgz
