+++ tests/bioinform_tests/gold_standard_results/bioinform_medium_goldstandard.tgz
