+++ tests/bioinform_tests/gold_standard_results/bioinform_short_goldstandard.tgz
