+++ tests/bioinform_tests/gold_standard_results/bioinform_tiny_goldstandard.tgz
