+++ tests/bioinform_tests/gold_standard_results/seek_basic_query_goldstandard.tgz
