+++ tests/bioinform_tests/gold_standard_results/seek_paramtest_goldstandard.tgz
