+++ tests/bioinform_tests/gold_standard_results/seek_querysize_goldstandard.tgz
