+++ tests/bioinform_tests/make_bioinform_config.sh
