+++ tests/bioinform_tests/pytools/struct_dict.py
