+++ tests/bioinform_tests/run_all.sh
