+++ tests/bioinform_tests/run_basic_query.sh
