+++ tests/bioinform_tests/run_bioinform.sh
