+++ tests/bioinform_tests/run_paramtest.sh
