+++ tests/bioinform_tests/run_querysize.sh
