+++ tests/example_add.cpp
@@ -1,0 +1,9 @@
+#include "gtest/gtest.h"
+#include "example_gtest.h"
+
+TEST(example, add)
+{
+    double res;
+    res = add_numbers(1.0, 2.0);
+    ASSERT_NEAR(res, 3.0, 1.0e-11);
+}
