+++ tests/example_subtract.cpp
@@ -1,0 +1,9 @@
+#include "gtest/gtest.h"
+#include "example_gtest.h"
+
+TEST(example, subtract)
+{
+    double res;
+    res = subtract_numbers(1.0, 2.0);
+    ASSERT_NEAR(res, -1.0, 1.0e-11);
+}
